+++ Source/SupportGLUT/Viewer/Screen.h
@@ -59,6 +59,9 @@ public:
     typedef kvs::glut::ScreenBase BaseClass;
     typedef kvs::Scene::ControlTarget ControlTarget;
 
+    static Screen* DownCast( kvs::ScreenBase* screen );
+    static const Screen* DownCast( const kvs::ScreenBase* screen );
+
 private:
     bool m_enable_default_paint_event; ///< flag for default paint event
     bool m_enable_default_resize_event; ///< flag for default resize event