+++ Source/SupportQt/Viewer/Screen.cpp
@@ -55,6 +55,30 @@ namespace qt
 
 /*===========================================================================*/
 /**
+ *  @brief  Returns the pointer to the qt screen downcasted from the screen base.
+ *  @param  screen [in] the screen base.
+ *  @return pointer to the qt screen
+ */
+/*===========================================================================*/
+Screen* Screen::DownCast( kvs::ScreenBase* screen )
+{
+    return dynamic_cast<Screen*>( screen );
+}
+
+/*===========================================================================*/
+/**
+ *  @brief  Returns the const pointer to the qt screen downcasted from the screen base.
+ *  @param  screen [in] the screen base.
+ *  @return const pointer to the qt screen
+ */
+/*===========================================================================*/
+const Screen* Screen::DownCast( const kvs::ScreenBase* screen )
+{
+    return dynamic_cast<Screen*>( const_cast<kvs::ScreenBase*>( screen ) );
+}
+
+/*===========================================================================*/
+/**
  *  @brief  Constructs a new Screen class.
  */
 /*===========================================================================*/