+++ Source/SupportQt/Viewer/ScreenBase.h
@@ -44,6 +44,10 @@ class ScreenBase : public QGLWidget, public kvs::ScreenBase
 
     typedef kvs::ScreenBase BaseClass;
 
+public:
+    static ScreenBase* DownCast( kvs::ScreenBase* screen );
+    static const ScreenBase* DownCast( const kvs::ScreenBase* screen );
+
 private:
 
     int m_id; ///< window ID