+++ CHANGELOG.md
@@ -64,6 +64,10 @@
 + kvs::TableObject::addColumn( array, min_value, max_value, label )
 + kvs::KeyEvent::modifiers
 + kvs::KeyEvent::setModifiers
++ kvs::glut::ScreenBase::DownCast
++ kvs::glut::Screen::DownCast
++ kvs::qt::ScreenBase::DownCast
++ kvs::qt::Screen::DownCast
 
 **Added new examples**
 + Example/Visualization/ScatterPlotMatrixRenderer